+++ packages/microsoft-security-adapter/src/definitions/deploy/shared/deploy.ts
@@ -1,0 +1,55 @@
+/*
+ *                      Copyright 2024 Salto Labs Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import _ from 'lodash'
+import { definitions, deployment } from '@salto-io/adapter-components'
+import { AdditionalAction, ClientOptions } from '../../types'
+import { DeployCustomDefinitions } from './types'
+import { createEntraCustomizations } from '../entra/deploy'
+import { omitReadOnlyFields } from './utils'
+
+const createCustomizations = (): DeployCustomDefinitions => {
+  const standardRequestDefinitions = deployment.helpers.createStandardDeployDefinitions<
+    AdditionalAction,
+    ClientOptions
+  >({})
+
+  return _.merge(standardRequestDefinitions, createEntraCustomizations())
+}
+
+export const createDeployDefinitions = (): definitions.deploy.DeployApiDefinitions<never, ClientOptions> => ({
+  instances: {
+    default: {
+      requestsByAction: {
+        default: {
+          request: {
+            context: deployment.helpers.DEFAULT_CONTEXT,
+            transformation: {
+              adjust: omitReadOnlyFields,
+            },
+          },
+          condition: {
+            transformForCheck: {
+              adjust: omitReadOnlyFields,
+            },
+          },
+        },
+        customizations: {},
+      },
+      changeGroupId: deployment.grouping.selfGroup,
+    },
+    customizations: createCustomizations(),
+  },
+})
