+++ packages/microsoft-security-adapter/src/definitions/fetch/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { createFetchDefinitions } from './entra/fetch'
+export { createFetchDefinitions } from './shared/fetch'
