+++ packages/microsoft-security-adapter/src/definitions/fetch/shared/defaults.ts
