+++ packages/microsoft-security-adapter/src/definitions/fetch/shared/fetch.ts
@@ -1,0 +1,43 @@
+/*
+ *                      Copyright 2024 Salto Labs Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { definitions } from '@salto-io/adapter-components'
+import { UserFetchConfig } from '../../../config'
+import { Options } from '../../types'
+import { DEFAULT_FIELD_CUSTOMIZATIONS, DEFAULT_ID_PARTS } from './defaults'
+import { createEntraCustomizations } from '../entra/fetch'
+
+const createCustomizations = (): Record<string, definitions.fetch.InstanceFetchApiDefinitions<Options>> =>
+  createEntraCustomizations()
+
+export const createFetchDefinitions = (
+  _fetchConfig: UserFetchConfig,
+): definitions.fetch.FetchApiDefinitions<Options> => ({
+  instances: {
+    default: {
+      resource: {
+        serviceIDFields: ['id'],
+      },
+      element: {
+        topLevel: {
+          elemID: { parts: DEFAULT_ID_PARTS },
+        },
+        fieldCustomizations: DEFAULT_FIELD_CUSTOMIZATIONS,
+      },
+    },
+    customizations: createCustomizations(),
+  },
+})
