+++ packages/microsoft-security-adapter/src/definitions/references/references.ts
@@ -19,7 +19,7 @@ import { ODATA_TYPE_FIELD_NACL_CASE, entraConstants } from '../../constants'
 import { Options } from '../types'
 import { REFERENCE_RULES } from './entra_reference_rules'
 
-const { SUPPORTED_DIRECTORY_OBJECT_ODATA_TYPE_NAME_TO_TYPE_NAME } = entraConstants
+const { SUPPORTED_DIRECTORY_OBJECTS_ODATA_TYPE_NAME_TO_TYPE_NAME } = entraConstants
 
 export const REFERENCES: definitions.ApiDefinitions<Options>['references'] = {
   rules: REFERENCE_RULES,
@@ -35,7 +35,7 @@ export const REFERENCES: definitions.ApiDefinitions<Options>['references'] = {
     ODataType: referenceUtils.neighborContextGetter({
       contextFieldName: ODATA_TYPE_FIELD_NACL_CASE,
       getLookUpName: async ({ ref }) => ref.elemID.name,
-      contextValueMapper: refType => SUPPORTED_DIRECTORY_OBJECT_ODATA_TYPE_NAME_TO_TYPE_NAME[refType],
+      contextValueMapper: refType => SUPPORTED_DIRECTORY_OBJECTS_ODATA_TYPE_NAME_TO_TYPE_NAME[refType],
     }),
   },
 }
